requirements.txt
GitPython==3.1.43
Jinja2==3.1.4
MarkupSafe==2.1.5
MedPy==0.5.2
PyGObject==3.42.1
PyJWT==2.3.0
PyYAML==6.0.2
Pygments==2.18.0
SecretStorage==3.3.1
Send2Trash==1.8.3
SimpleITK==2.4.0
albucore==0.0.14
albumentations==1.4.14
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
autocommand==2.2.2
babel==2.16.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==3.4.8
dbus-python==1.2.18
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
docker-pycreds==0.4.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fqdn==1.5.1
fsspec==2024.6.1
gitdb==4.0.11
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.2
idna==3.8
imageio==2.35.1
importlib-metadata==4.6.4
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
iotop==0.6
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.1
jeepney==0.7.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema-specifications==2023.12.1
jsonschema==4.23.0
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
matplotlib-inline==0.1.7
mistune==3.0.2
monai==1.3.2
more-itertools==10.3.0
more-itertools==8.10.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
notebook_shim==0.2.4
numpy==2.1.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-python-headless==4.10.0.84
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pip==22.0.2
pip==24.2
platformdirs==4.2.2
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.26.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.1
sentry-sdk==2.13.0
setproctitle==1.3.3
setuptools==59.6.0
setuptools==74.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.2
terminado==0.18.1
tifffile==2024.8.28
tinycss2==1.3.0
tomli==2.0.1
tomli==2.0.1
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
typeguard==4.3.0
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wadllib==1.3.6
wandb==0.17.8
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.37.1
wheel==0.43.0
wheel==0.44.0
zipp==1.0.0
zipp==3.19.2